requirements.txt
pymongo==4.6.0
python-dotenv==1.0.0
tqdm==4.66.1
psutil==5.9.6
colorama==0.4.6